requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
coloredlogs==15.0.1
Django==4.2.2
flatbuffers==23.5.26
humanfriendly==10.0
idna==3.4
imageio==2.31.1
importlib-metadata==6.7.0
lazy-loader==0.2
llvmlite==0.40.1rc1
mpmath==1.3.0
networkx==3.1
numba==0.57.0
numpy==1.24.3
onnxruntime==1.15.1
opencv-python-headless==4.7.0.72
packaging==23.1
Pillow==9.5.0
platformdirs==3.7.0
pooch==1.7.0
protobuf==4.23.3
PyMatting==1.1.8
PyWavelets==1.4.1
rembg==2.0.46
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
sqlparse==0.4.4
sympy==1.12
tifffile==2023.4.12
tqdm==4.65.0
typing-extensions==4.6.3
urllib3==2.0.3
zipp==3.15.0
